Add VV to Subject Preview
