+++ packages/lib-react-components/package.json
@@ -42,6 +42,7 @@
     "@visx/responsive": "~3.12.0",
     "@visx/shape": "~3.12.0",
     "@zooniverse/panoptes-js": "~0.5.0",
+    "@zooniverse/subject-viewers": "0.1.0",
     "cuid": "~3.0.0",
     "formik": "~2.4.0",
     "i18next": "~24.2.2",
