+++ packages/lib-react-components/src/Media/Media.stories.js
@@ -28,6 +28,8 @@ const VIDEO_URL =
   'https://static.zooniverse.org/fem-assets/home-video.mp4'
 const TEXT_URL =
   'https://panoptes-uploads.zooniverse.org/subject_location/f5506d1c-a0e9-4aba-a418-6a6c46a7731a.txt'
+const VOLUMETRIC_URL = 
+  'https://panoptes-uploads-staging.zooniverse.org/subject_location/3c56e6a7-be27-4ab3-93a7-c41a3c867baf.json'
 
 export function Image() {
   return (
@@ -96,6 +98,19 @@ export function TextMedia() {
       <Media
         alt='A text file'
         src={TEXT_URL}
+        width={270}
+      />
+    </Box>
+  )
+}
+
+export function VolumetricMedia() {
+  return (
+    <Box>
+      <Text>Width set as 270</Text>
+      <Media
+        alt='A volumetric file'
+        src={VOLUMETRIC_URL}
         width={270}
       />
     </Box>
