+++ packages/lib-react-components/src/Media/components/Data/components/Volumetric/Volumetric.js
@@ -1,0 +1,18 @@
+import { lazy, Suspense } from 'react'
+
+const VolumetricViewerComponent = lazy(() => import('@zooniverse/subject-viewers/VolumetricViewer'))
+
+function VolumetricViewer({ data }) {
+  const config = {
+    subject: {
+      subjectJSON: data
+    },
+    view: 'preview'
+  }
+
+  return <Suspense fallback={<p>Loading Volumetric Viewer...</p>}>
+    <VolumetricViewerComponent {...config} />
+  </Suspense>
+}
+
+export default VolumetricViewer
