+++ packages/lib-react-components/src/Media/components/Data/components/Volumetric/index.js
@@ -1,0 +1,1 @@
+export { default } from './Volumetric'
