+++ packages/lib-react-components/src/Media/components/Data/components/index.js
@@ -1,1 +1,2 @@
 export { default as DataSeriesPlot } from './DataSeriesPlot'
+export { default as Volumetric } from './Volumetric'
