+++ packages/lib-react-components/src/types/JSONData/Volumetric.js
@@ -1,0 +1,8 @@
+import { types } from 'mobx-state-tree'
+
+const Volumetric = types.model('Volumetric', {
+  type: types.string,
+  data: types.string
+})
+
+export default Volumetric
